v6.5.1

+++ released/MyBox/src/main/java/mara/mybox/bufferedimage/CombineTools.java
@@ -25,12 +25,12 @@ public class CombineTools {
         try {
             int x = imageCombine.getMarginsValue();
             int y = imageCombine.getMarginsValue();
-            int imageWidth;
-            int imageHeight;
-            int totalWidth = 0;
-            int totalHeight = 0;
-            int maxHeight = 0;
-            int minHeight = Integer.MAX_VALUE;
+            double imageWidth;
+            double imageHeight;
+            double totalWidth = 0;
+            double totalHeight = 0;
+            double maxHeight = 0;
+            double minHeight = Integer.MAX_VALUE;
             int sizeType = imageCombine.getSizeType();
             if (isPart) {
                 y = 0;
@@ -57,7 +57,9 @@ public class CombineTools {
             for (ImageInformation imageInfo : images) {
                 imageWidth = imageInfo.getWidth();
                 imageHeight = imageInfo.getHeight();
-                if (sizeType == ImageCombine.CombineSizeType.KeepSize || sizeType == ImageCombine.CombineSizeType.TotalWidth || sizeType == ImageCombine.CombineSizeType.TotalHeight) {
+                if (sizeType == ImageCombine.CombineSizeType.KeepSize
+                        || sizeType == ImageCombine.CombineSizeType.TotalWidth
+                        || sizeType == ImageCombine.CombineSizeType.TotalHeight) {
                 } else if (sizeType == ImageCombine.CombineSizeType.EachWidth) {
                     imageHeight = (imageHeight * imageCombine.getEachWidthValue()) / imageWidth;
                     imageWidth = imageCombine.getEachWidthValue();
@@ -73,8 +75,8 @@ public class CombineTools {
                 }
                 xs.add(x);
                 ys.add(y);
-                widths.add(imageWidth);
-                heights.add(imageHeight);
+                widths.add((int) imageWidth);
+                heights.add((int) imageHeight);
                 x += imageWidth + imageCombine.getIntervalValue();
                 if (imageHeight > totalHeight) {
                     totalHeight = imageHeight;
@@ -84,9 +86,11 @@ public class CombineTools {
             if (!isPart) {
                 totalHeight += 2 * imageCombine.getMarginsValue();
             }
-            Image newImage = combineImages(images, totalWidth, totalHeight,
-                    FxColorTools.toAwtColor(imageCombine.getBgColor()), xs, ys, widths, heights, imageCombine.getTotalWidthValue(), imageCombine.getTotalHeightValue(),
-                    careTotal && (sizeType == ImageCombine.CombineSizeType.TotalWidth), careTotal && (sizeType == ImageCombine.CombineSizeType.TotalHeight));
+            Image newImage = combineImages(images, (int) totalWidth, (int) totalHeight,
+                    FxColorTools.toAwtColor(imageCombine.getBgColor()), xs, ys, widths, heights,
+                    imageCombine.getTotalWidthValue(), imageCombine.getTotalHeightValue(),
+                    careTotal && (sizeType == ImageCombine.CombineSizeType.TotalWidth),
+                    careTotal && (sizeType == ImageCombine.CombineSizeType.TotalHeight));
             return newImage;
         } catch (Exception e) {
             MyBoxLog.error(e.toString());
@@ -135,12 +139,12 @@ public class CombineTools {
         try {
             int x = imageCombine.getMarginsValue();
             int y = x;
-            int imageWidth;
-            int imageHeight;
-            int totalWidth = 0;
-            int totalHeight = 0;
-            int maxWidth = 0;
-            int minWidth = Integer.MAX_VALUE;
+            double imageWidth;
+            double imageHeight;
+            double totalWidth = 0;
+            double totalHeight = 0;
+            double maxWidth = 0;
+            double minWidth = Integer.MAX_VALUE;
             int sizeType = imageCombine.getSizeType();
             if (sizeType == ImageCombine.CombineSizeType.AlignAsBigger) {
                 for (ImageInformation imageInfo : imageInfos) {
@@ -184,8 +188,8 @@ public class CombineTools {
                 }
                 xs.add(x);
                 ys.add(y);
-                widths.add(imageWidth);
-                heights.add(imageHeight);
+                widths.add((int) imageWidth);
+                heights.add((int) imageHeight);
                 x = imageCombine.getMarginsValue();
                 y += imageHeight + imageCombine.getIntervalValue();
                 if (imageWidth > totalWidth) {
@@ -194,7 +198,7 @@ public class CombineTools {
             }
             totalWidth += 2 * imageCombine.getMarginsValue();
             totalHeight = y + imageCombine.getMarginsValue() - imageCombine.getIntervalValue();
-            Image newImage = combineImages(imageInfos, totalWidth, totalHeight,
+            Image newImage = combineImages(imageInfos, (int) totalWidth, (int) totalHeight,
                     FxColorTools.toAwtColor(imageCombine.getBgColor()), xs, ys, widths, heights, imageCombine.getTotalWidthValue(), imageCombine.getTotalHeightValue(),
                     careTotal && (sizeType == ImageCombine.CombineSizeType.TotalWidth), careTotal && (sizeType == ImageCombine.CombineSizeType.TotalHeight));
             return newImage;
